Fix systematic warning reduction: S125, S1450, S3881, S3626, S1066 warnings resolved

+++ FlinkDotNet/FlinkDotNet.Storage.FileSystem.Tests/FileSystemSnapshotReaderTests.cs
@@ -33,23 +33,10 @@ namespace FlinkDotNet.Storage.FileSystem.Tests
 
 
         {
-
-
-            if (disposing)
-
-
-            {
-
-
-                if (Directory.Exists(_testDirectory))
+            if (disposing && Directory.Exists(_testDirectory))
             {
                 Directory.Delete(_testDirectory, recursive: true);
-
-
             }
-
-
-        }
         }
 
         private async Task<string> CreateTestSnapshot()
